Change player prefs handling class from being static to be a service that can be injected and replaced/mocked
