+++ Assets/Scripts/Services/ILocalStorageService.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 533c546c06dbd4522b038dd61f5e467c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
